ファイル関係の関数をcommonに移動
