+++ packages/common/src/index.ts
@@ -16,3 +16,5 @@ export const config = {
         )
     )
 };
+
+export * from "./files.js";
