+++ packages/common/tsconfig.json
@@ -7,6 +7,8 @@
     "strict": true,
     "skipLibCheck": true,
     "rootDir": "src",
-    "outDir": "dist"
-  }
+    "outDir": "dist",
+    "declaration": true
+  },
+  "include": ["src/**/*.ts"]
 }
